requirements.txt
certifi==2021.10.8
click==8.0.3
Flask==2.0.2
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.3
keras==2.7.0
MarkupSafe==2.0.1
numpy==1.21.5
Pillow==8.4.0
python-dateutil==2.8.2
python-dotenv==0.19.2
tensorflow
pytz==2021.3
six==1.16.0
Werkzeug==2.0.2
